requirements.txt
telethon==1.34.0
python-telegram-bot==20.7
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
gspread==5.12.0
google-auth==2.27.0
aiohttp==3.9.0
psutil==5.9.8
